--- a/modules/index-service/src/main/java/org/opencastproject/index/service/impl/index/series/Series.java
+++ b/modules/index-service/src/main/java/org/opencastproject/index/service/impl/index/series/Series.java
@@ -169,6 +169,22 @@ public class Series implements IndexObject {
   }
 
   /**
+   * Create an unmarshaller for series
+   * @return an unmarshaller for series
+   * @throws IOException
+   */
+  public static Unmarshaller createUnmarshaller() throws IOException {
+    try {
+      if (context == null) {
+        createJAXBContext();
+      }
+      return context.createUnmarshaller();
+    } catch (JAXBException e) {
+      throw new IOException(e.getLinkedException() != null ? e.getLinkedException() : e);
+    }
+  }
+
+  /**
    * Returns the series identifier.
    *
    * @return the identifier
@@ -523,15 +539,15 @@ public class Series implements IndexObject {
    *
    * @param xml
    *          the input stream
+   * @param unmarshaller the unmarshaller to use
    * @return the deserialized recording event
    * @throws IOException
    */
-  public static Series valueOf(InputStream xml) throws IOException {
+  public static Series valueOf(InputStream xml, Unmarshaller unmarshaller) throws IOException {
     try {
       if (context == null) {
         createJAXBContext();
       }
-      Unmarshaller unmarshaller = context.createUnmarshaller();
       return unmarshaller.unmarshal(new StreamSource(xml), Series.class).getValue();
     } catch (JAXBException e) {
       throw new IOException(e.getLinkedException() != null ? e.getLinkedException() : e);