The following requests are quite slow when asking for a large number of objects:

- GET /admin-ng/events
- GET /api/events
- GET /admin-ng/series
- GET /api/series

An analysis revealed that most of the time is spent creating JAXB Unmarshaller objects for each single item. This can be sped up by reusing the same Unmarshaller between events/series.

Note that this is not thread-safe, of course, but we’re inside a single thread while deserializing, so this should be fine.

We saw the following performance improvement when displaying 100 items per page:

Events table: From approx. 850ms down to approx. 150ms
Series table: From approx. 900ms down to approx. 250ms

So the re-use of JAXB Unmarshaller objects results in a speedup of roughly 4-5.

This work is sponsored by SWITCH.