@pmiddend 2018-10-25
Sorry for screwing up the tests, I will fix it immediately.